--- src/modules/App/App.reducer.test.ts
@@ -1,10 +1,0 @@
-import * as SUT from './App.reducer'
-
-describe('modules/App/App.reducer', () => {
-  describe('resetAppData', () => {
-    it('should clear the app state', () => {
-      const action = SUT.resetAppData()
-      expect(action).toBeTruthy()
-    })
-  })
-})
