--- src/modules/App/App.reducer.ts
@@ -1,5 +1,0 @@
-import { createAction } from '@reduxjs/toolkit'
-
-const name = '[App]'
-
-export const resetAppData = createAction(`${name} clear app data`)
