+++ src/modules/App/types.ts
@@ -1,0 +1,3 @@
+export enum AppActions {
+  resetApp = '[App] Clear App Reducers',
+}
